requirements.txt
paho-mqtt 
tinydb
requests 
numpy 
opencv-python-headless
keras
tensorflow
keras-preprocessing 
# https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_x86_64.whl
PyTurbiJPEG